Workbench modules added for content moderation
Certain things about the workbench suit of modules do not work
very intutively for the end user (eg: my sections tab that only
confuses and does nothing, drafts by a user not listed once a
reviewer has edit and therefore becomes the author of latest revision
of a content etc.) and therefore we would need to create issues on DO and
and submit patches for them.

Some of the tweaking around of workbench to suit our needs can also be
done using a custom 'adc_tweaks' module.

references code4aap/drupal-aap-campaign#14